\name{describe.gamma.survey}
\alias{describe.gamma.survey}
\title{Describe a gamma...}
\usage{describe.gamma.survey(...)}
\description{Describe a \code{gamma.survey} Citation to Zelig}
\value{a list to be processed by \code{as.description}}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{...}{ignored parameters}}
