\name{describe.logit.survey}
\alias{describe.logit.survey}
\title{Describe a logit...}
\usage{describe.logit.survey(...)}
\description{Describe a \code{logit.survey} Citation to Zelig}
\value{a list to be processed by \code{as.description}}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{...}{ignored parameters}}
