\name{describe.normal.survey}
\alias{describe.normal.survey}
\title{Describe a normal...}
\usage{describe.normal.survey(...)}
\description{Describe a \code{normal.survey} Citation to Zelig}
\value{a list to be processed by \code{as.description}}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{...}{ignored parameters}}
