\name{describe.poisson.survey}
\alias{describe.poisson.survey}
\title{Describe a poisson...}
\usage{describe.poisson.survey(...)}
\description{Describe a \code{poisson.survey} Citation to Zelig}
\value{a list to be processed by \code{as.description}}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{...}{ignored parameters}}
