\name{describe.probit.survey}
\alias{describe.probit.survey}
\title{Describe a probit...}
\usage{describe.probit.survey(...)}
\description{Describe a \code{probit.survey} Citation to Zelig}
\value{a list to be processed by \code{as.description}}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{...}{ignored parameters}}
