\name{param.logit.survey}
\alias{param.logit.survey}
\title{Param Method for the logit...}
\usage{\method{param}{logit.survey}(obj, num=1000, ...)}
\description{Param Method for the \code{logit.survey} Zelig Model}
\note{This method is used internally by the \code{survey.zelig} package}
\value{a list to be cast as a \code{parameters} object}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{obj}{a \code{zelig} object}
\item{num}{an integer specifying the number of simulations to sample}
\item{...}{ignored parameters}}
