\name{plot.sim.gamma.survey}
\alias{plot.sim.gamma.survey}
\title{Plot Simulated Quantities of Interest for probit...}
\usage{\method{plot}{sim.gamma.survey}(x, xlab="", user.par=FALSE, ...)}
\description{Plot Simulated Quantities of Interest for \code{probit.survey}}
\value{This function is used for its side-effects}
\arguments{\item{x}{A \code{qi} object containing simulated quantities of interest}
\item{xlab}{Label for the x-axis}
\item{user.par}{User-specified graphical parameters}
\item{...}{ignored parameters}}
