\name{survey.Zelig-package}
\alias{survey.Zelig-package}
\alias{survey.Zelig}
\docType{package}
\title{
  Survey-Weighted Models for Zelig
}
\description{
  Implements survey-weighted models using the Zelig API.
}
\details{
\tabular{ll}{
Package: \tab survey.Zelig\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-01-10\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
}
\author{
Matthew Owen, Olivia Lau, Kosuke Imai, Gary King

Maintainer: mowen@iq.harvard.edu
}
\references{

}
\keyword{ package }
\keyword{ survey }
\keyword{ Zelig }

\seealso{
}
\examples{

}
