\name{survey.zelig-package}
\title{Survey-Weighted Regression Models for Zelig...}
\description{Survey-Weighted Regression Models for Zelig}
\details{\tabular{ll}{
Package: \tab survey.zelig\cr
Version: \tab 1.0-1\cr
Date: \tab 2011-05-05\cr
Depends: \tab Zelig, survey\cr
License: \tab GPL version 2 or newer\cr
URL: \tab http://gking.harvard.edu/zelig\cr
}

Add-on pack for Zelig, containing ordinal models:
\tabular{ll}{
gamma.survey: \tab Survey-Weighted Gamma Regression \cr
logit.survey: \tab Survey-Weighted Logit Regression \cr
normal.survey: \tab Survey-Weighted Normal Regression \cr
poisson.survey: \tab Survey-Weighted Poisson Regression \cr
probit.survey: \tab Survey-Weighted Probit Regression \cr
}}
\alias{survey.zelig}
\docType{package}
\author{Matt Owen \email{mowen@iq.harvard.edu}, Kosuke Imai, Olivia Lau,
and Gary King 

Maintainer: Matt Owen \email{mowen@iq.harvard.edu}}
\keyword{package}
